refactor(model-servers): keep port parsing local
